pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["mkorrabi <korrabimadjid111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.3"
pyspark = "3.4.0"
kafka-python-ng = "2.2.2"
pandas = "2.1.2"
flightradarapi = "^1.3.31"
findspark = "^2.0.1"
pytz = "^2024.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
papermill = "^2.6.0"
ipykernel = "^6.29.5"
notebook = "^7.2.1"
jupyterlab = "^4.2.4"
pyarrow = "^17.0.0"
fastparquet = "^2024.5.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
